Refactor Imaging Tests
