+++ framework/test/Volo.Abp.Imaging.Abstractions.Tests/Volo/Abp/Imaging/AbpImagingAbstractionsTestBase.cs
@@ -1,0 +1,11 @@
+﻿using Volo.Abp.Testing;
+
+namespace Volo.Abp.Imaging;
+
+public abstract class AbpImagingAbstractionsTestBase : AbpIntegratedTest<AbpImagingAbstractionsTestModule>
+{
+    protected override void SetAbpApplicationCreationOptions(AbpApplicationCreationOptions options)
+    {
+        options.UseAutofac();
+    }
+}
