+++ framework/test/Volo.Abp.Imaging.Abstractions.Tests/Volo/Abp/Imaging/AbpImagingAbstractionsTestModule.cs
