+++ framework/test/Volo.Abp.Imaging.AspNetCore.Tests/Volo.Abp.Imaging.AspNetCore.Tests.abppkg.json
@@ -1,0 +1,3 @@
+{
+  "role": "lib.test"
+}
