+++ framework/test/Volo.Abp.Imaging.AspNetCore.Tests/Volo/Abp/Imaging/AbpImagingAspNetCoreTestBase.cs
@@ -1,0 +1,11 @@
+﻿using Volo.Abp.Testing;
+
+namespace Volo.Abp.Imaging;
+
+public abstract class AbpImagingAspNetCoreTestBase : AbpIntegratedTest<AbpImagingAspNetCoreTestModule>
+{
+    protected override void SetAbpApplicationCreationOptions(AbpApplicationCreationOptions options)
+    {
+        options.UseAutofac();
+    }
+}
