+++ framework/test/Volo.Abp.Imaging.AspNetCore.Tests/Volo/Abp/Imaging/AbpImagingAspNetCoreTestModule.cs
@@ -1,0 +1,14 @@
+﻿using Volo.Abp.Autofac;
+using Volo.Abp.Modularity;
+
+namespace Volo.Abp.Imaging;
+
+[DependsOn(
+    typeof(AbpAutofacModule),
+    typeof(AbpImagingAspNetCoreModule),
+    typeof(AbpTestBaseModule)
+)]
+public class AbpImagingAspNetCoreTestModule : AbpModule
+{
+
+}
