+++ framework/test/Volo.Abp.Imaging.ImageSharp.Tests/Volo.Abp.Imaging.ImageSharp.Tests.abppkg.json
@@ -1,0 +1,3 @@
+{
+  "role": "lib.test"
+}
