+++ framework/test/Volo.Abp.Imaging.ImageSharp.Tests/Volo/Abp/Imaging/AbpImagingImageSharpTestBase.cs
