+++ framework/test/Volo.Abp.Imaging.ImageSharp.Tests/Volo/Abp/Imaging/AbpImagingImageSharpTestModule.cs
