+++ framework/test/Volo.Abp.Imaging.MagickNet.Tests/Volo.Abp.Imaging.MagickNet.Tests.abppkg.json
@@ -1,0 +1,3 @@
+{
+  "role": "lib.test"
+}
