+++ framework/test/Volo.Abp.Imaging.MagickNet.Tests/Volo/Abp/Imaging/AbpImagingMagickNetTestBase.cs
