+++ framework/test/Volo.Abp.Imaging.MagickNet.Tests/Volo/Abp/Imaging/AbpImagingMagickNetTestModule.cs
